+++ scripts/assertTsNode.ts
@@ -19,15 +19,9 @@ declare global {
 	}
 }
 
-let detectTSNode = false;
-
 try {
-	if (process[REGISTER_INSTANCE]) {
-		detectTSNode = true;
-	}
-} catch {}
-
-if (!detectTSNode) {
+	process[REGISTER_INSTANCE];
+} catch {
 	throw new EvalError(
 		"Couldn't detect the ts-node environment. Did you import this script as a module by mistake?"
 	);