+++ src/events/interactionCreate.ts
@@ -129,7 +129,8 @@ async function handleCommandInteraction(
 			};
 
 			try {
-				return await command.execute(messageContextMenuCommandContext);
+				await command.execute(messageContextMenuCommandContext);
+				return;
 			} catch (error) {
 				await sendErrorMessage(interaction, error);
 				return;
@@ -156,7 +157,8 @@ async function handleCommandInteraction(
 			};
 
 			try {
-				return await command.execute(userContextMenuCommandContext);
+				await command.execute(userContextMenuCommandContext);
+				return;
 			} catch (error) {
 				await sendErrorMessage(interaction, error);
 				return;
@@ -164,7 +166,8 @@ async function handleCommandInteraction(
 		}
 
 		try {
-			return await command.execute(context);
+			await command.execute(context);
+			return;
 		} catch (error) {
 			await sendErrorMessage(interaction, error);
 			return;
@@ -174,14 +177,16 @@ async function handleCommandInteraction(
 	if (context.source === 'dm') {
 		// No guild found
 		logger.debug(`Command '${command.info.name}' requires guild information, but none was found.`);
-		return await context.reply({
+		await context.reply({
 			content: "Can't do that here",
 			ephemeral: true,
 		});
+		return;
 	}
 
 	try {
-		return await command.execute(context);
+		await command.execute(context);
+		return;
 	} catch (error) {
 		await sendErrorMessage(interaction, error);
 	}
@@ -204,7 +209,8 @@ async function handleAutocompleteInteraction(interaction: AutocompleteInteractio
 				`Received request to execute autocomplete handler for unknown command named '${interaction.commandName}'`
 			);
 			// Return no results
-			return await interaction.respond([]);
+			await interaction.respond([]);
+			return;
 		}
 
 		// Command must be a chat-input command
@@ -213,7 +219,8 @@ async function handleAutocompleteInteraction(interaction: AutocompleteInteractio
 				`Received an autocomplete request for command '${command.info.name}'. This command must be of type 'ChatInput', but was found instead to be of a different type (${command.type}).`
 			);
 			// Return no results
-			return await interaction.respond([]);
+			await interaction.respond([]);
+			return;
 		}
 
 		// Command must have an autocomplete handler
@@ -222,7 +229,8 @@ async function handleAutocompleteInteraction(interaction: AutocompleteInteractio
 				`Received an autocomplete request for command '${command.info.name}'. This command must have an autocomplete handler method, but none was found.`
 			);
 			// Return no results
-			return await interaction.respond([]);
+			await interaction.respond([]);
+			return;
 		}
 
 		logger.debug(`Calling autocomplete handler for command '${command.info.name}'`);
@@ -328,7 +336,8 @@ async function handleButtonInteraction(
 		channelId: interaction.channelId,
 	};
 	try {
-		return await button.execute(buttonContext);
+		await button.execute(buttonContext);
+		return;
 	} catch (error) {
 		await sendErrorMessage(interaction, error);
 	}