+++ src/evilHangman/evilHangmanAsciiArt.ts
@@ -3,7 +3,8 @@ import fs from 'node:fs/promises';
 const ART_PATH = './res/hangmen.txt';
 const FRAME_SEPARATOR = 'æ';
 const framesPromise = (async (): Promise<ReadonlyArray<string>> => {
-	const allFrames = (await fs.readFile(ART_PATH)).toString();
+	const file = await fs.readFile(ART_PATH);
+	const allFrames = file.toString();
 	return allFrames.split(FRAME_SEPARATOR);
 })();