+++ src/helpers/guards/isNonEmptyArray.test.ts
@@ -14,7 +14,7 @@ describe('Nonempty array', () => {
 		${4}
 		${128}
 	`('returns true for an array of $length item(s)', ({ length }: { length: number }) => {
-		const array = new Array<number>(length).fill(8);
+		const array = Array.from({ length }).fill(8);
 		expect(array.length).toBe(length); // sanity check
 		expect(isNonEmptyArray(array)).toBe(true);
 	});