+++ src/helpers/positionsOfUriInText.ts
@@ -12,17 +12,17 @@ export function positionsOfUriInText(str: string): NonEmptyArray<Range> | null {
 		/https?:\/\/(www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*)/gu;
 
 	let results: NonEmptyArray<Range> | null = null;
-	let match: RegExpExecArray | null = null;
+	let match: RegExpExecArray | null;
 
 	while ((match = uri.exec(str))) {
 		const range: Range = {
 			start: match.index,
-			end: match.index + (match[0]?.length ?? 0),
+			end: match.index + match[0].length,
 		};
-		if (!results) {
-			results = [range];
-		} else {
+		if (results) {
 			results.push(range);
+		} else {
+			results = [range];
 		}
 	}